refactor: move root to separate file
